requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.1





























































































































































































































































































































































































































































































































































































































































































































httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygsheets==2.0.6
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
telebot==0.0.5
psycopg2-binary==2.9.6
